[ Refactor ] A code change that neither fixes a bug (UI)
