+++ src/components/common/SortBlock.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const SortBlock = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
