--- src/components/common/index.jsx
@@ -1,1 +1,0 @@
-import React from 'react';
