--- src/components/layout/DanggeunImage.js
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-import danggeun from '../../images/ico/ico_logo_72x72.png';
-
-export const Danggeun = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  margin-left: 145px;
-
-  width: 72px;
-  height: 72px;
-
-  background-image: url(${danggeun});
-  background-size: cover;
-`;
