+++ src/index.js
@@ -6,14 +6,18 @@ import { BrowserRouter } from 'react-router-dom';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import App from './App';
 import rootReducer from './modules';
+import ResetStyles from './resetStyle/ResetStyles';
 
 const store = createStore(rootReducer, composeWithDevTools());
 
 ReactDOM.render(
-  <BrowserRouter>
-    <Provider store={store}>
-      <App />
-    </Provider>
-  </BrowserRouter>,
+  <>
+    <ResetStyles />
+    <BrowserRouter>
+      <Provider store={store}>
+        <App />
+      </Provider>
+    </BrowserRouter>
+  </>,
   document.getElementById('root'),
 );
