--- src/pages/Gps.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-import MenuBar from '../components/common/MenuBar';
-
-const GpsBlock = styled.div`
-  margin-top: 100px;
-  text-align: center;
-  font-size: 20px;
-`;
-
-const Gps = () => (
-  <>
-    <GpsBlock>성재님이 만드실 페이지입니다.</GpsBlock>
-  </>
-);
-
-export default Gps;
