+++ src/pages/MyDanggeun.js
@@ -1,5 +1,3 @@
-// 나의 당근 화면
-
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { useSelector } from 'react-redux';
@@ -8,8 +6,6 @@ import {
   MyDanggeunBlock,
   ProfileLine,
   ProfileName,
-  ListLine,
-  TopImageBlock,
   ProfileUrl,
   SaleBlock,
   PurchaseBlock,
@@ -17,28 +13,16 @@ import {
   LogOut,
   WatchBlock,
   TextBlock,
-  MyAddressBlock,
-  AddressAuthBlock,
-  KeywordBlock,
-  CollectBlock,
-  ListLineTwo,
-  TownWrite,
-  TownComment,
-  TownTheme,
-  ListLineThree,
-  CeoMenu,
-  ListLineFour,
-  FriendBlock,
-  DanggeunShare,
+  Line,
   ProfileImage,
-  NoticeBlock,
-  QnABlock,
   ListBlock,
-  AppSet,
   Block,
-  LogoutLine,
-} from '../styles/MyDanggeunStyle';
+  Ptag,
+  TownBlock,
+  DefaultBlock,
+} from '../components/layout/MyDanggeunStyle';
 import firebase from '../utils/api/fbInstance';
+import MenuBar from '../components/common/MenuBar';
 
 const MyDanggeun = () => {
   const { name } = useSelector(({ profilename }) => ({
@@ -56,10 +40,7 @@ const MyDanggeun = () => {
   return (
     <>
       <TopBlock />
-      <MyDanggeunBlock>
-        <TopImageBlock />
-        나의 당근
-      </MyDanggeunBlock>
+      <MyDanggeunBlock>나의 당근</MyDanggeunBlock>
 
       <ProfileImage />
 
@@ -68,90 +49,99 @@ const MyDanggeun = () => {
       </Link>
       <ProfileName>
         {name}
-        <span>안양6동</span>
+        <div>안양6동</div>
       </ProfileName>
       <ProfileLine />
 
       <Block>
-        <SaleBlock />
-        <PurchaseBlock />
-        <WatchBlock />
+        <DefaultBlock>
+          <SaleBlock />
+        </DefaultBlock>
+        <DefaultBlock>
+          <PurchaseBlock />
+        </DefaultBlock>
+        <DefaultBlock>
+          <WatchBlock />
+        </DefaultBlock>
       </Block>
 
       <TextBlock>
-        <p>판매내역</p>
-        <p>구매내역</p>
-        <p>관심목록</p>
+        <Ptag>판매내역</Ptag>
+        <Ptag>구매내역</Ptag>
+        <Ptag>관심목록</Ptag>
       </TextBlock>
 
-      <ListLine />
+      <Line />
 
       <ListBlock>
-        <MyAddressBlock>
-          <p>내 동네 설정</p>
-        </MyAddressBlock>
-        <AddressAuthBlock>
-          <p>동네 인증하기</p>
-        </AddressAuthBlock>
-        <KeywordBlock>
-          <p>키워드 알림</p>
-        </KeywordBlock>
-        <CollectBlock>
-          <p>모아보기</p>
-        </CollectBlock>
+        <TownBlock>
+          <Ptag>내 동네 설정</Ptag>
+        </TownBlock>
+        <TownBlock>
+          <Ptag>동네 인증하기</Ptag>
+        </TownBlock>
+        <TownBlock>
+          <Ptag>키워드 알림</Ptag>
+        </TownBlock>
+        <TownBlock>
+          <Ptag>모아보기</Ptag>
+        </TownBlock>
       </ListBlock>
 
-      <ListLineTwo />
+      <Line />
 
       <ListBlock>
-        <TownWrite>
-          <p>동네생활 글</p>
-        </TownWrite>
-        <TownComment>
-          <p>동네생활 댓글</p>
-        </TownComment>
-        <TownTheme>
-          <p>동네생활 주제 목록</p>
-        </TownTheme>
+        <TownBlock>
+          <Ptag>동네생활 글</Ptag>
+        </TownBlock>
+        <TownBlock>
+          <Ptag>동네생활 댓글</Ptag>
+        </TownBlock>
+        <TownBlock>
+          <Ptag>동네생활 주제 목록</Ptag>
+        </TownBlock>
       </ListBlock>
 
-      <ListLineThree />
+      <Line />
 
       <ListBlock>
-        <CeoMenu>
-          <p>사장님 메뉴</p>
-        </CeoMenu>
+        <TownBlock>
+          <Ptag>사장님 메뉴</Ptag>
+        </TownBlock>
       </ListBlock>
 
-      <ListLineFour />
+      <Line />
 
       <ListBlock>
-        <FriendBlock>
-          <p>친구초대</p>
-        </FriendBlock>
-        <DanggeunShare>
-          <p>당근마켓 공유</p>
-        </DanggeunShare>
-        <NoticeBlock>
-          <p>공지사항</p>
-        </NoticeBlock>
-        <QnABlock>
-          <p>자주 묻는 질문</p>
-        </QnABlock>
-        <AppSet>
-          <p>앱 설정</p>
-        </AppSet>
+        <TownBlock>
+          <Ptag>친구초대</Ptag>
+        </TownBlock>
+        <TownBlock>
+          <Ptag>당근마켓 공유</Ptag>
+        </TownBlock>
+        <TownBlock>
+          <Ptag>공지사항</Ptag>
+        </TownBlock>
+        <TownBlock>
+          <Ptag>자주 묻는 질문</Ptag>
+        </TownBlock>
+        <TownBlock>
+          <Ptag>앱 설정</Ptag>
+        </TownBlock>
       </ListBlock>
-      <LogoutLine />
+
+      <Line />
 
       <ListBlock>
         <Link to="/login">
           <LogOut onClick={Logout}>
-            <p>로그아웃</p>
+            <Ptag>로그아웃</Ptag>
           </LogOut>
         </Link>
         <EmptyBlock />
       </ListBlock>
+
+      <MenuBar />
     </>
   );
 };
