--- src/styles/Layout.js
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const Layout = styled.div`
-  padding: 0 18px;
-  height: 100%;
-`;
